pyproject.toml
[project]
name = "doc2md"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "litellm>=1.74.14",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
packages = ["doc2md"]
include-package-data = true


[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
